requirements.txt
# requirements.txt
biopython>=1.81
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
streamlit>=1.22.0  # optional
numpy>=1.23.0